Fix t/08_readmember_record_sep.t for Win32
The test was being skipped in Win32 because of test failures.

The test failures were caused by A::Z reading/writing members in binary
mode, i.e. without any newline conversion. This has to be taken into
account in the test, by:
- Converting "\n" into "\r\n" in the data file prior to writing to the
  zip file, if running in Win32.
- Converting "\n" of the expected strings returned by getline() into
  "\r\n" before checking, if running in Win32.

These changes were made to the test file and all tests pass on Win32.
On other platforms the changes have no effect, and should pass the tests
also.