Prevent from traversing symlinks and parent directories when extracting
If an attacker-supplied archive contains symbolic links and files that
referes to the symbolic links in their path components, the user can
be tricked into overwriting any arbitrary file.

The same issue is with archives whose members refer to a parent
directory (..) in their path components.

This patch fixes it by aborting an extraction (extractTree(),
extractMember(), extractMemberWithoutPaths()) in those cases by not
traversing the dangerous paths and returning AZ_ERORR instead.

However, if a user supplies a local file name, the security checks are
not performed. This is based on the assumption that a user knows
what's on his local file system.

CVE-2018-10860
https://bugzilla.redhat.com/show_bug.cgi?id=1591449