Accept filenames that are stringifying objects
Sometimes a filename may be an object which stringifies to its path. As
such, it should work as a string anywhere that strings are. Modules
implementing such objects include Path::Tiny and Path::Class; they enable
filenames to be derived from components, other filenames, and so on.

Unfortunately Archive::Zip was, when passed such a filename object to the
->read method, trying to treat the filename as a file-handle, and failing
with the surprising error message:

  Can't locate object method "opened" via package "Path::Tiny" at
  /whatever/Archive-Zip/lib/Archive/Zip/Archive.pm line 585.

The attached patch fixes this, and provides a test for the behaviour.

It isn't even documented that the ->read method can accept a file-handle.
(The ->readFromFileHandle method is provided for that.) However, since it
was working, the patch is careful to ensure that that still works, and
contains a test for it: only an object that stringifies is treated as a
filename; other objects continue to be treated as file-handles.