fix: allow symlinks and `..` elements in `dest`
That is, when extracting a tree into a specified destination directory,
do not bail out because:

    1. The destination directory specification contains `..` elements,
    2. The destination directory is a symlink, or
    3. One or more directories in the parentage of the destination
       directory are symlinks.

This makes `extractTree` behave more like `extractMember` when the
latter is passed an explicit extraction target path.  With these
changes, `Archive::Zip` continues to guard against the security threat
of path traversals, while not placing undue restrictions on users'
ability to select extraction destination directories.